requirements.txt
click==7.1.2
Flask==1.1.2
grpcio==1.36.1
grpcio-tools==1.36.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
protobuf==3.15.6
six==1.15.0
Werkzeug==1.0.1
